Improve performance by using modified `kinet2pcb`
Found out that `hierplace` is taking majority of time when creating PCB
out of netlist. This is not needed for our use-case because we place
footprints with `kbplacer`. For details see [1].
This should help with server stability - we currently run on lowest tier
linode machine and bigger layouts were taking a lot of time to process.

[1] https://github.com/adamws/kinet2pcb/commit/f3cb8dc56492c87897edecbe50b8b26cf65feade